+++ config/initializers/spotlight_initializer.rb
@@ -5,8 +5,7 @@ Spotlight::Engine.config.upload_fields = [
   OpenStruct.new(field_name: 'date', solr_field: %w(spotlight_upload_date_tesim date_sort), label: 'Date')
 ]
 Spotlight::Engine.config.default_contact_email = Settings.default_contact_email
-Spotlight::Engine.config.external_resources_partials += ['dor_harvester/form']
-Spotlight::Engine.config.external_resources_partials += ['bibliography_resources/form'] if FeatureFlags.new.bibliography_resource?
+Spotlight::Engine.config.external_resources_partials += ['dor_harvester/form', 'bibliography_resources/form']
 
 Spotlight::Resources::Upload.document_builder_class = ::UploadSolrDocumentBuilder