+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Adds an "Expand all" link to nested related items in the full metadata display #960
 ### Changed
+- BibTeX upload form is now only shown for Exhibits in which it is explicitly enabled through a feature flag #946
 ### Deprecated
 ### Removed
 ### Fixed